// fe/src/lib/api/generated/model/cSMCC2Value.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CSMCC2Value = typeof CSMCC2Value[keyof typeof CSMCC2Value];


export const CSMCC2Value = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;